requirements.txt
absl-py==0.7.0
appnope==0.1.0
asn1crypto==0.24.0
astor==0.7.1
astroid==2.1.0
attrs==18.2.0
Augmentor==0.2.3
Automat==0.7.0
backcall==0.1.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
configparser==3.7.3
constantly==15.1.0
cryptography==2.4.2
cssselect==1.0.3
cycler==0.10.0
decorator==4.3.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.1.7
djangorestframework==3.9.1
et-xmlfile==1.0.1
future==0.17.1
gast==0.2.2
gcloud==0.18.3
googleapis-common-protos==1.5.8
grpcio==1.18.0
gunicorn==19.9.0
httplib2==0.12.1
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
isort==4.3.4
itsdangerous==1.1.0
jdcal==1.4
jedi==0.13.2
Jinja2==2.10
jupyter-client==5.2.4
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
lxml==4.3.0
Markdown==3.0.1
MarkupSafe==1.1.1
matplotlib==3.0.2
mccabe==0.6.1
mock==2.0.0
numpy==1.16.0
oauth2client==4.1.3
opencv-python==4.0.0.21
openpyxl==2.5.14
pandas==0.23.4
parsel==1.5.1
parso==0.3.1
pbr==5.1.2
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
prompt-toolkit==2.0.7
protobuf==3.6.1
psutil==5.5.1
ptyprocess==0.6.0
pyasn1==0.4.5
pyasn1-modules==0.2.3
pycparser==2.19
PyDispatcher==2.0.5
Pygments==2.3.1
PyHamcrest==1.9.0
pylint==2.2.2
pyOpenSSL==18.0.0
pyparsing==2.3.1
PyQt5==5.11.3
PyQt5-sip==4.19.13
python-dateutil==2.7.5
pytz==2018.9
PyYAML==3.13
pyzmq==17.1.2
queuelib==1.5.0
requests==2.21.0
rsa==4.0
scikit-learn==0.20.2
scipy==1.2.1
Scrapy==1.5.1
scrapy-splash==0.7.2
service-identity==18.1.0
six==1.12.0
static3==0.7.0
termcolor==1.1.0
tornado==5.1.1
tqdm==4.29.1
traitlets==4.3.2
Twisted==18.9.0
urllib3==1.24.1
w3lib==1.20.0
wcwidth==0.1.7
Werkzeug==0.14.1
wrapt==1.11.1
zope.interface==4.6.0
psycopg2==2.7.3.1
